+++ server/service/endpoint_users.go
@@ -34,29 +34,6 @@ func makeGetSessionUserEndpoint(svc fleet.Service) endpoint.Endpoint {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-// Change Password
-////////////////////////////////////////////////////////////////////////////////
-
-type changePasswordRequest struct {
-	OldPassword string `json:"old_password"`
-	NewPassword string `json:"new_password"`
-}
-
-type changePasswordResponse struct {
-	Err error `json:"error,omitempty"`
-}
-
-func (r changePasswordResponse) error() error { return r.Err }
-
-func makeChangePasswordEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(changePasswordRequest)
-		err := svc.ChangePassword(ctx, req.OldPassword, req.NewPassword)
-		return changePasswordResponse{Err: err}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
 // Reset Password
 ////////////////////////////////////////////////////////////////////////////////
